int to float

+++ detectie/train.py
@@ -62,7 +62,7 @@ def main(args):
 if __name__ == '__main__':
     parser = ArgumentParser()
     parser.add_argument('--num_workers', type=int, default=12)
-    parser.add_argument('--weight_decay', type=int, default=0)
+    parser.add_argument('--weight_decay', type=float, default=0)
     parser.add_argument('--momentum', type=float, default=0.95)
     parser.add_argument('--lr', type=float, default=1e-3)
     parser.add_argument('--batch_size', type=int, default=1)
